requirements.txt
python-dotenv
langchain-openai
langchain-core
tqdm
openai